refactor: extract Issue functions to github/issues.py (#139)
